Using one package for the api instead of many
Instead of making small mini packages everywhere. Using one big package
to get the api results
